Performance counters back to 64 bits

+++ main.cpp
@@ -159,8 +159,8 @@ int main(int, char**)
 
 
     // Delta Time
-	uint32_t dt_NOW = SDL_GetPerformanceCounter();
-    uint32_t dt_LAST = 0;
+	uint64_t dt_NOW = SDL_GetPerformanceCounter();
+    uint64_t dt_LAST = 0;
 	float deltaTime = 0.f;
 
     // Main loop
